Strip Messaging::Configuration Static Behavior.
The nature of having ``Messaging::Configuration`` as a singleton and
thusly static artificially limits the configurability of multiple
users of this library, should one want to use connections to multiple
different brokers in the same process.

The only price that a user of this would have to pay is by providing
a skeletal implementation of ``Messaging::Client.configuration`` in
whatever class they use the mixin with.  As illustrated in the
``examples`` directory, this needn't add superfluousness to the
interface behaviors of this library.

A negative byproduct hereof is the breaking of API compatibility with
previous versions, but that shouldn't be a reason to not incorporate
on the grounds that it actually improves testability, grokability,
and compartmentalization of concerns for future users.

I can confirm that ``make``, ``make test``, and
``bundle exec examples/run`` all succeed happily in my local
development environment.